refactor(watchlist-exclusions): split page into table, columns, and toolbar
